Revert "Refactor types to match API"
This reverts commit b14d9b45d913b430b339f949bc279885bf87ad1d.

+++ interface.go
@@ -13,7 +13,7 @@ type TimeZone struct {
 
 // Measure is used to contain a value and its units of measure
 type Measure struct {
-	Value    float64 `json:"Value"`
+	Value    float32 `json:"Value"`
 	Unit     string  `json:"Unit"`
 	UnitType int     `json:"UnitType"`
 }
